Move Modal back into ExpandButton
